files are moved around. make sure to reroute
