+++ src/routers/signupRouter.js
@@ -2,11 +2,11 @@ const express = require('express');
 const app = express();
 const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
-mongoose.Promise = global.Promise;
+//mongoose.Promise = global.Promise;
 
 const {User} = require('./models');
 
-app.use(bodyParser.json());
+//app.use(bodyParser.json());
 const router = express.Router();
 
 router.get('/', (req, res) => {
