--- userRouter.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-router.get('/', (req, res) => {
-  res.sendFile(__dirname + '/user-page.html');
-});
-
-module.exports = router;
